# Improve the performance of ActiveSupport::MessageEncryptor