### Summary

Following up on the work I did in #42919, this introduces the same logic to `ActiveSupport::MessageEncryptor`.

By calculating the position of the IV and auth tags, instead of using `String#split`, we increase our resistance to invalid messages, those that don't feature the separator. The new code is not sensitive to the size of the input whenever it doesn't feature the separator.

Besides the above, I'm making sure that the auth tag is 16 bytes, OpenSSL bindings' current default. A change to that default would break things for us, namely making it impossible to decrypt previously encrypted messages.

<!-- Provide a general description of the code changes in your pull
request... were there any bugs you had fixed? If so, mention them. If
these bugs have open GitHub issues, be sure to tag them here as well,
to keep the conversation linked together. -->

### Other Information

#### Benchmarks

**4 kB**

Attempting to decrypt an invalid message (no separator).

Currently:
```
Warming up --------------------------------------
 #decrypt_and_verify     5.403k i/100ms
   #encrypt_and_sign     6.674k i/100ms
Calculating -------------------------------------
 #decrypt_and_verify     54.628k (± 4.6%) i/s -    545.703k in  10.013536s
   #encrypt_and_sign     66.277k (± 5.3%) i/s -    667.400k in  10.100746s
````

This PR:
```
Warming up --------------------------------------
 #decrypt_and_verify     7.098k i/100ms
   #encrypt_and_sign     7.577k i/100ms
Calculating -------------------------------------
 #decrypt_and_verify     74.996k (± 4.5%) i/s -    752.388k in  10.054574s
   #encrypt_and_sign     73.336k (± 4.0%) i/s -    734.969k in  10.039760s
```

A 1.37x speedup for `#decrypt_and_verify`, and a 1.10x speedup for `#encrypt_and_sign`.

---

Attempting to decrypt a valid message.

Currently:
```
Warming up --------------------------------------
 #decrypt_and_verify     5.366k i/100ms
   #encrypt_and_sign     6.464k i/100ms
Calculating -------------------------------------
 #decrypt_and_verify     52.541k (± 2.9%) i/s -    525.868k in  10.017648s
   #encrypt_and_sign     68.079k (± 3.8%) i/s -    685.184k in  10.080207s
```

This PR:
```
Warming up --------------------------------------
 #decrypt_and_verify     5.787k i/100ms
   #encrypt_and_sign     7.331k i/100ms
Calculating -------------------------------------
 #decrypt_and_verify     57.518k (± 5.6%) i/s -    578.700k in  10.094993s
   #encrypt_and_sign     72.982k (± 4.4%) i/s -    733.100k in  10.066154s
```

A 1.09x speedup for `#decrypt_and_verify`, and a 1.07x speedup for `#encrypt_and_sign`.

**4 MB**

Attempting to decrypt an invalid message (no separator).

Currently:
```
Warming up --------------------------------------
 #decrypt_and_verify    21.000  i/100ms
   #encrypt_and_sign    10.000  i/100ms
Calculating -------------------------------------
 #decrypt_and_verify    239.920  (± 1.7%) i/s -      2.415k in  10.069074s
   #encrypt_and_sign    113.399  (± 2.6%) i/s -      1.140k in  10.058523s
```

This PR:
```
Warming up --------------------------------------
 #decrypt_and_verify     7.257k i/100ms
   #encrypt_and_sign    12.000  i/100ms
Calculating -------------------------------------
 #decrypt_and_verify     78.076k (± 1.9%) i/s -    783.756k in  10.042234s
   #encrypt_and_sign    132.422  (± 3.8%) i/s -      1.332k in  10.074115s
```

A **326x** speedup for `#decrypt_and_verify`, and a 1.16x speedup for `#encrypt_and_sign`.

---

Attempting to decrypt a valid message.

Currently:
```
Warming up --------------------------------------
 #decrypt_and_verify    10.000  i/100ms
   #encrypt_and_sign    11.000  i/100ms
Calculating -------------------------------------
 #decrypt_and_verify    100.046  (± 4.0%) i/s -      1.000k in  10.013520s
   #encrypt_and_sign    122.601  (± 4.1%) i/s -      1.232k in  10.068093s
```

This PR:
```
Warming up --------------------------------------
 #decrypt_and_verify    12.000  i/100ms
   #encrypt_and_sign    14.000  i/100ms
Calculating -------------------------------------
 #decrypt_and_verify    152.493  (± 3.3%) i/s -      1.524k in  10.005288s
   #encrypt_and_sign    137.735  (± 1.5%) i/s -      1.386k in  10.065756s
```

A 1.52x speedup for `#decrypt_and_verify`, and a 1.12x speedup for `#encrypt_and_sign`.

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here. This could include
benchmarks, or other information.

If you are updating any of the CHANGELOG files or are asked to update the
CHANGELOG files by reviewers, please add the CHANGELOG entry at the top of the file.

Finally, if your pull request affects documentation or any non-code
changes, guidelines for those changes are [available
here](https://edgeguides.rubyonrails.org/contributing_to_ruby_on_rails.html#contributing-to-the-rails-documentation)

Thanks for contributing to Rails! -->

<!--
Note: Please avoid making *Draft* pull requests, as they still send
notifications to everyone watching the Rails repo.
Create a pull request when it is ready for review and feedback
from the Rails team :).
-->